`EuiOverlayMask` converted to TS
