--- src/components/overlay_mask/index.js
@@ -1,3 +1,0 @@
-export {
-  EuiOverlayMask,
-} from './overlay_mask';
