+++ src/components/overlay_mask/index.ts
@@ -1,0 +1,1 @@
+export { EuiOverlayMask } from './overlay_mask';
